requirements_new.txt
# Automatically generated by https://github.com/damnever/pigar.

adan-pytorch==0.1.0
dadaptation==3.1
einops
h5py
numpy
timm
#torch
torchinfo
wandb
tqdm
zarr
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
mpi4py
#if run into errors like "src/mpi4py.MPI.c:6:10: fatal error: Python.h: No such file or directory", check https://stackoverflow.com/questions/21530577/fatal-error-python-h-no-such-file-or-directory
scipy
matplotlib
netCDF4
git+https://github.com/sandialabs/exodusii.git